Move tryFetchJson to new file
